Indel filter
The indel filter script also filters out low quality sites.
I also updated the bash script, to call the indel filter and rm the need for user input to overwrite files in the output folder - since it caused issues with looping the script directly from the command line. I guess I could bring it back if the loop was done within the script.