fasta iupac
Validating fasta I found more bugs in the pipeline, which has been modified in the bash script. The bcftools consensus was outputting IUPAC ambiguity code when the site was homozygous for the alternate allele. I don't ask for iupac anymore and now replace the alleles directly in the fasta file "by hand" in the iupac_fasta.py script. Preliminary validation seems to be working, I will need to validate all the files again tomorrow after they are generated.